+++ scorched/tests/test_functional.py
@@ -14,6 +14,12 @@ class TestUtils(unittest.TestCase):
             self.datajson = f.read()
             self.docs = json.loads(self.datajson)
 
+    def tearDown(self):
+        dsn = os.environ.get("SOLR_URL",
+                             "http://localhost:8983/solr")
+        si = SolrInterface(dsn)
+        si.delete_all()
+
     @scorched.testing.skip_unless_solr
     def test_query(self):
         dsn = os.environ.get("SOLR_URL",
@@ -23,7 +29,16 @@ class TestUtils(unittest.TestCase):
         si.commit()
         res = si.query(genre_s="fantasy").execute()
         self.assertEqual(res.result.numFound, 3)
-        si.delete_all()
+        # delete
+        si.delete_by_ids(res[0]['id'])
+        si.commit()
+        res = si.query(genre_s="fantasy").execute()
+        self.assertEqual(res.result.numFound, 2)
+        # TODO rollback
+        # we see a rollback in solr log but entry is still deleted
+        #si.rollback()
+        #res = si.query(genre_s="fantasy").execute()
+        #self.assertEqual(res.result.numFound, 3)
 
     @scorched.testing.skip_unless_solr
     def test_facet_query(self):
@@ -45,7 +60,6 @@ class TestUtils(unittest.TestCase):
                           'facet_dates': {},
                           'facet_queries': {},
                           'facet_pivot': ()})
-        si.delete_all()
 
     @scorched.testing.skip_unless_solr
     def test_filter_query(self):
@@ -59,7 +73,6 @@ class TestUtils(unittest.TestCase):
         self.assertEqual(res.result.numFound, 1)
         self.assertEqual([x['name'] for x in res.result.docs],
                          [u'The Lightning Thief'])
-        si.delete_all()
 
     @scorched.testing.skip_unless_solr
     def test_edismax_query(self):
@@ -73,4 +86,44 @@ class TestUtils(unittest.TestCase):
         self.assertEqual(res.result.numFound, 1)
         self.assertEqual([x['name'] for x in res.result.docs],
                          [u'The Lightning Thief'])
-        si.delete_all()
+
+    @scorched.testing.skip_unless_solr
+    def test_mlt_component_query(self):
+        dsn = os.environ.get("SOLR_URL",
+                             "http://localhost:8983/solr")
+        si = SolrInterface(dsn)
+        si.add(self.docs)
+        si.commit()
+        res = si.query(id="978-0641723445").mlt(
+            "genre_s", mintf=1, mindf=1).execute()
+        # query shows only one
+        self.assertEqual(res.result.numFound, 1)
+        # but in more like this we get two
+        self.assertEqual(len(res.more_like_these["978-0641723445"].docs), 2)
+        self.assertEqual([x['author'] for x in res.more_like_these[
+            "978-0641723445"].docs], [u'Rick Riordan', u'Jostein Gaarder'])
+
+
+class TestMltHandler(unittest.TestCase):
+
+    def setUp(self):
+        file = os.path.join(os.path.dirname(__file__), "dumps",
+                            "books.json")
+        with open(file) as f:
+            self.datajson = f.read()
+            self.docs = json.loads(self.datajson)
+
+    @scorched.testing.skip_unless_solr
+    def test_mlt(self):
+        dsn = os.environ.get("SOLR_URL",
+                             "http://localhost:8983/solr")
+        si = SolrInterface(dsn)
+        si.add(self.docs)
+        si.commit()
+        res = si.mlt_query("genre_s",
+                           interestingTerms="details", mintf=1, mindf=1
+                           ).query(id="978-0641723445").execute()
+        self.assertEqual(res.result.numFound, 2)
+        self.assertEqual(res.interesting_terms, [u'genre_s:fantasy', 1.0])
+        self.assertEqual([x['author'] for x in res.result.docs],
+                         [u'Rick Riordan', u'Jostein Gaarder'])